requirements.txt
streamlit
genai
Pillow
